+++ .gitignore
@@ -187,5 +187,14 @@ pyrightconfig.json
 /benchmark/ex4/mf6rtm
 /benchmark/ex5/mf6rtm
 *.png
+
+#Python pixi environments 
 pixi.lock
 pixi.toml
+
+#Test output files
+/autotest/test01/*
+/autotest/test02/*
+/autotest/test03/*
+/autotest/test04/*
+/autotest/test05/*
